requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.21.0
safetensors>=0.3.0